feat(midi-io): wrap MIDI-CI, UMP-Stream & Flex-metadata negotiators in ECS
The tutti-layer MIDI 2.0 negotiators (CiResponder/CiInitiator,
EndpointInquiry, JrStamper) were pure state machines with no ECS home.
Add the outbound + resource half of the ECS wiring, mirroring the
existing device/routing sub-plugin shape:

- negotiation.rs — MidiNegotiationPlugin: CiRes (responder + initiator),
  EndpointDiscoveryRes; Start{Ci,Endpoint}Discovery request messages drive
  outbound probes onto the bus; Inbound{CiMessage,EndpointReply} feed the
  negotiators and raise {CiDevice,Endpoint}Discovered results.
- metadata.rs — MidiMetadataPlugin: BroadcastFlexMetadata (chord/key/text)
  emits Flex Data in-band on the bus; JrStamperRes homes the JR-stamp config.
- MidiBus::broadcast_ci fragments a CiMessage over the SysEx7 pipeline.

Inbound decode stays a stub seam (the hardware input path drops SysEx),
so Inbound{CiMessage,EndpointReply} have no in-crate producer yet — the
injection point a later inbound-decode pass or a test writes into. A
headless-app test drives ci_ingest_system end to end to prove the seam is
live, not dead code. Also drops a redundant u16::from in jr_timestamp_value.

Verified: tutti-midi-io 22 tests, tutti-midi-runtime 95 tests green;
clippy -D warnings clean; bevy-tutti (app workspace) builds against the
new umbrella re-exports.

+++ crates/midi/tutti-midi-types/src/ump/utility.rs
@@ -32,7 +32,7 @@ impl MidiEvent {
         use midi2::utility::Utility;
         use midi2::UmpMessage;
         match UmpMessage::try_from(self.data_words()).ok()? {
-            UmpMessage::Utility(Utility::Timestamp(m)) => Some(u16::from(m.time_data())),
+            UmpMessage::Utility(Utility::Timestamp(m)) => Some(m.time_data()),
             _ => None,
         }
     }
